refactor: post annotation adapted, transactional refactored, README.md contents added
