+++ community/src/main/java/org/community/annotation/CurrentPost.java
@@ -1,10 +1,4 @@
 package org.community.annotation;
 
-import java.lang.annotation.*;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
 public @interface CurrentPost {
-
 }
